Common Leave modulee
